refactor: module renames to match contained classes
