+++ apis/python/src/tiledbsoma/experiment.py
@@ -6,7 +6,7 @@ from .collection import CollectionBase
 from .dataframe import DataFrame
 from .experiment_query import AxisQuery, ExperimentAxisQuery
 from .measurement import Measurement
-from .soma_session_context import TileDBSessionContext
+from .tiledb_session_context import TileDBSessionContext
 from .tiledb_object import TileDBObject
