+++ apis/python/tests/test_io.py
@@ -6,7 +6,7 @@ from scipy import sparse as sp
 import tiledbsoma as soma
 import tiledbsoma.io as somaio
 from tiledbsoma import TileDBCreateOptions
-from tiledbsoma.soma_session_context import TileDBSessionContext
+from tiledbsoma.tiledb_session_context import TileDBSessionContext
 
 
 @pytest.fixture
